QR code generator/requirements.txt
Pillow==12.1.0
qrcode==8.2
